Change data type `np.float` -> `float`
`np.float` is removed in newer version of numpy.

+++ ptemcee/sampler.py
@@ -267,12 +267,12 @@ class Sampler(object):
         if betas is not None:
             self._betas = betas
 
-        self.nswap = np.zeros(self.ntemps, dtype=np.float)
-        self.nswap_accepted = np.zeros(self.ntemps, dtype=np.float)
+        self.nswap = np.zeros(self.ntemps, dtype=float)
+        self.nswap_accepted = np.zeros(self.ntemps, dtype=float)
 
-        self.nprop = np.zeros((self.ntemps, self.nwalkers), dtype=np.float)
+        self.nprop = np.zeros((self.ntemps, self.nwalkers), dtype=float)
         self.nprop_accepted = np.zeros(
-            (self.ntemps, self.nwalkers), dtype=np.float)
+            (self.ntemps, self.nwalkers), dtype=float)
 
         if random is not None:
             self._random = random
@@ -510,9 +510,9 @@ class Sampler(object):
         mapf = map if self.pool is None else self.pool.map
         results = list(mapf(self._likeprior, ps.reshape((-1, self.dim))))
 
-        logl = np.fromiter((r[0] for r in results), np.float,
+        logl = np.fromiter((r[0] for r in results), float,
                            count=len(results)).reshape((self.ntemps, -1))
-        logp = np.fromiter((r[1] for r in results), np.float,
+        logp = np.fromiter((r[1] for r in results), float,
                            count=len(results)).reshape((self.ntemps, -1))
 
         return logl, logp
